requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.4
datasets==3.6.0
deprecation==2.1.0
dill==0.3.8
fastapi==0.115.13
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.3.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.0
google-auth==2.40.3
googleapis-common-protos==1.70.0
gotrue==2.12.0
greenlet==3.2.3
grpcio==1.73.0
grpcio-status==1.73.0
hpack==4.1.0
huggingface-hub==0.33.0
idna==3.10
iniconfig==2.1.0
instagrapi==2.1.5
instaloader==4.14.1
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.1.5
langchain-text-splitters==0.3.8
langsmith==0.3.45
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.5
numpy==2.3.0
orjson==3.10.18
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pluggy==1.6.0
postgrest==1.0.2
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.26.1
pypdfium2==4.30.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.0
pytest-mock==3.14.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
realtime==2.4.3
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
starlette==0.46.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.3
supafunc==0.9.4
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.1
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.52.4
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.4.0
uvicorn==0.34.3
websocket-client==1.8.0
websockets==14.2
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
youtube-search-python==1.6.6
zstandard==0.23.0
fastapi
python-dotenv
requests
langchain
langchain-core
langchain-google-genai
langchain-community